requirements.txt
xmltodict==0.13.0
fastapi==0.88.0
pydantic==1.10.4
python-multipart==0.0.5
prometheus_client==0.15.0
confluent-kafka==1.9.2
jsonschema==4.17.3
ujson==5.6.0
python-logging-loki==0.3.1
prometheus_fastapi_instrumentator==5.9.1
opentelemetry-exporter-otlp==1.14.0
opentelemetry-instrumentation-fastapi==0.36b0
opentelemetry-instrumentation-logging==0.36b0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-confluent-kafka==0.36b0
uvicorn==0.20.0
gunicorn==20.1.0